+++ src/jtl/Connector/Shopware/Mapper/CustomerOrder.php
@@ -63,7 +63,7 @@ class CustomerOrder extends DataMapper
             ->leftJoin('customer_shipping.attribute', 'customer_shipping_attribute')
             ->leftJoin('customer.debit', 'debit')
             ->leftJoin('orders.attribute', 'attribute')
-            ->leftJoin('orders.details', 'details')
+            ->join('orders.details', 'details')
             ->leftJoin('details.tax', 'tax')
             ->leftJoin('orders.billing', 'billing')
             ->leftJoin('orders.shipping', 'shipping')